Move to useHook
